Create separate file for fill tests
